add avail light client submitter for the batcher
