+++ packages/batcher/runtime/src/types.ts
@@ -1,8 +1,8 @@
 import type { RequestHandler } from 'express';
-import type BatchedTransactionPoster from '@paima/batcher-transaction-poster';
+import type { BatchedTransactionPoster } from '@paima/batcher-transaction-poster';
 import type GameInputValidator from '@paima/batcher-game-input-validator';
 import type { Pool } from 'pg';
-import type { EthersEvmProvider } from '@paima/providers';
+import type { EthersEvmProvider, PolkadotProvider } from '@paima/providers';
 
 export interface BatcherRuntimeInitializer {
   initialize: (pool: Pool) => BatcherRuntime;
@@ -14,6 +14,7 @@ export interface BatcherRuntime {
   run: (
     gameInputValidator: GameInputValidator,
     BatchedTransactionPoster: BatchedTransactionPoster,
-    provider: EthersEvmProvider
+    provider: EthersEvmProvider | PolkadotProvider,
+    getCurrentBlock: () => Promise<number>
   ) => Promise<void>;
 }
